Fix an issue with hits from postings
If a document doesn't contain any hits for that term but some document
after it contains hits for that term then we'd use it's hits.  When you
combine it with hits from other terms which were on the document then
you'd get totally busted HitEnums where merging on position didn't also
merge on start/stop offset.