Fix highlighting on empty numeric fields
When there are no field contents and the field is not indexed with offsets
in postings or term vectors we used to build the elasticsearch provided
analyzer agaist empty string.  This is pretty crazy in the first place and
fails when you try to highlight all fields ("*") because _size is a hidden,
always there field that is usually empty.  And its numeric.  So for the
most part this behavior made it impossible to highlight "*".

Closes #9

Change-Id: I9c4b1c8ffe33f803be60b6b8aeaf0c95c8510194