cleanup checkstyle warnings on ...-core module
Javadoc warnings are left, I'm wondering if it make sense to try to fix them.
It will require a non trivial amount of energy, and doing javadoc after the
fact is likely to generate poor quality javadoc. I'm considering tagging all
the problematic classes to ignore the check, so we can ensure we write
javadoc on new classes and maybe slowly add it to existing classes.

Change-Id: I9b19e3a4220592f6d0b6571fde1a8668e3594022