Improve regex highlighting performance
The default forward algorithm, when matching with a regex like '.*foo', has a
very expensive failure case when provided a string that does not match the
regex (such as the tail of a document after the initial match). The forward
implementation requires length^2 state transitions to verify none of the
possible initial positions match.

Avoid this by first performing a backwards pass marking all valid start
positions of the regex. The forward pass can then lookup the next valid start
position and return a match directly. In this way the source is only scanned
once for each pass.

Limit this operations to detected leading wildcards (using a very
hacky method) since the additional bitset allocation is undesirable
work for 99.99% (only 800 out of 5M!) of production queries. For those
few that do need it though the improvement can be from minutes (and
highlighting doesn't time out!) to milliseconds.

Bug: T195491
Change-Id: I73c789c2d8d633486dd9ca1f9775ff1b2476b737