Add elastic 6.3.0 support
Was pretty straightforward except the fact that upstream moved
some analyzis components into the analysis-commons plugins.
Solution was to replicate their behaviors in a MockPlugin.
Far from ideal but these core plugins are not available as
maven dependencies, only as zip files.
This includes a potential fix for T178425 where similar
error started to appear in the test when using copy_to fields.

Bug: T198069
Bug: T178425
Change-Id: I7c8967b3e9e88f4d561696cd2be81e3f35cd8f54