Fix surrogate counting in regex matching
Count codepoints instead of characters when matching so that last
character matched in a row is not split into its surrogates.

Add lots of tests of 1- to 4-byte regexes matching in different
places within a word and having partial matches after full
matches to exercise all the codepoint counting code.

Add .DS_Store to .gitignore.

Bug: T404632
Change-Id: I365871a3f17b467ec500e3ebb849551f3fdc281a